requirements.txt
openai==1.30.5
groq==0.8.0
python-dotenv==1.0.1
streamlit==1.35.0
watchdog==4.0.1